+++ apps/client/src/components/hub/chatForm/ChatFormEditor.js
@@ -23,6 +23,19 @@ export default function ChatFormEditor() {
 
   return (
     <>
+      <div className="w-full py-[10px] flex justify-center rounded-lg bg-(--color-accent-light)">
+        <p className="!font-light">
+          Please read the documentation before you start:{" "}
+          <a
+            href="https://oleksandr-ch.notion.site/Programmable-Chat-feature-1ffe3b41e4ae804da619f63b706e7263"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="!font-normal cursor-pointer underline text-(--color-accent-dark)"
+          >
+            documentation
+          </a>
+        </p>
+      </div>
       <ChatEditorCode />
       <ChatEditorLogs logs={compilerLogs} setLogs={setCompilerLogs} />
       <div className="h-[61px] px-5 py-2 flex-shrink-0 flex justify-between items-center rounded-2xl bg-[var(--color-hover-light)]">