+++ apps/client/src/components/message/LastMessage.js
@@ -24,12 +24,13 @@ export default function LastMessage({
     );
   }
 
-  const { attachments, body } = message;
+  const { attachments, body } = message || {};
   const lastAtt = attachments?.slice(-1)[0];
 
   const lastMessageText = (text, att) => {
     if (text) return text;
-    if (att?.file_name) return getFileType(att.file_name);
+    if (att?.file_name || att?.file_content_type)
+      return getFileType(att.file_name, att.file_content_type);
     return "";
   };