+++ apps/client/src/lib/downloadManager.js
@@ -16,9 +16,7 @@ export default class DownloadManager {
           mUpdate[mid] = { _id: mid, attachments: [] };
         }
         mUpdate[mid].attachments.push({
-          file_id: attachments[fileId].file_id,
-          file_name: attachments[fileId].file_name,
-          file_blur_hash: attachments[fileId].file_blur_hash,
+          ...attachments[fileId],
           file_url: urls[fileId],
         });
       }