+++ apps/client/src/store/values/Conversations.js
@@ -139,7 +139,7 @@ export const conversations = createSlice({
       if (isRemove && !conv?.draft) return;
       const updateParams = { _id: cid };
 
-      if (conv.last_message.t > draft?.updated_at) {
+      if (conv.last_message?.t > draft?.updated_at) {
         if (draft) updateParams.draft = draft.message;
       } else if (conv.last_message) {
         updateParams.last_message = {
@@ -159,7 +159,9 @@ export const conversations = createSlice({
           updateParams.old_updated_at = null;
           updateParams.draft = null;
         } else {
-          updateParams.updated_at = draft.updated_at;
+          updateParams.updated_at = new Date(
+            draft.updated_at * 1000
+          ).toISOString();
           updateParams.old_updated_at = conv.updated_at;
           if (draft) updateParams.draft = draft.message;
         }