+++ apps/client/src/utils/media/get_file_type.js
@@ -1,13 +1,31 @@
-export default function getFileType(fileName) {
-  const fileExtension = fileName?.split(".").slice(-1)[0];
+const imageMimeTypes = [
+  "image/jpeg",
+  "image/jpg",
+  "image/png",
+  "image/gif",
+  "image/bmp",
+  "image/heic",
+];
+const videoMimeTypes = ["video/mp4", "video/webm", "video/quicktime"];
 
-  if (
-    ["jpeg", "jpg", "gif", "bmp", "png", "heic", "HEIC"].includes(fileExtension)
-  ) {
-    return "Image";
+const imageExtensions = ["jpeg", "jpg", "gif", "bmp", "png", "heic", "HEIC"];
+const videoExtensions = ["mp4", "webm", "mov"];
+
+export default function getFileType(fileName, mimeType) {
+  if (mimeType) {
+    if (imageMimeTypes.includes(mimeType)) {
+      return "Image";
+    }
+    if (videoMimeTypes.includes(mimeType)) {
+      return "Video";
+    }
   }
 
-  if (["mp4", "webm", "mov"].includes(fileExtension)) {
+  const fileExtension = fileName?.split(".").slice(-1)[0];
+  if (imageExtensions.includes(fileExtension)) {
+    return "Image";
+  }
+  if (videoExtensions.includes(fileExtension)) {
     return "Video";
   }